example/flask-gunicorn/requirements.txt
uptrace==1.10.0
flask==1.1.2
flask_sqlalchemy==2.5.1
opentelemetry-instrumentation-flask==0.29b0
opentelemetry-instrumentation-sqlalchemy==0.29b0
itsdangerous==1.1.0
gunicorn==20.0.4
